added conditional upload of deb files

+++ fog05-plugin-fdu-containerd/generate_deb.sh
@@ -2,6 +2,13 @@
 
 set -e
 
+if [[ -z "${DEPLOY}" ]]; then
+  UPLOAD=false
+else
+  UPLOAD=true
+fi
+
+
 UBUNTU="ubuntu:bionic"
 DEBIAN="debian:10-slim"
 
@@ -34,10 +41,11 @@ docker cp build-ctd:/root/build/fog05-plugin-fdu-containerd_${VERSION}-1_amd64.d
 
 docker container rm --force build-ctd
 
-
-set +x
-echo $KEY  | base64 --decode > key
-chmod 0600 key
-scp -o StrictHostKeyChecking=no -i ./key ../fog05-plugin-fdu-containerd_${VERSION}-1_amd64.deb $USER@$SERVER:$DEPLOYDIR/fos/deb/bionic/amd64/fog05-plugin-fdu-containerd-1_amd64.deb
-rm key
-set -x
+if ["$UPLOAD" = true ]; then
+    set +x
+    echo $KEY  | base64 --decode > key
+    chmod 0600 key
+    scp -o StrictHostKeyChecking=no -i ./key ../fog05-plugin-fdu-containerd_${VERSION}-1_amd64.deb $USER@$SERVER:$DEPLOYDIR/fos/deb/bionic/amd64/fog05-plugin-fdu-containerd-1_amd64.deb
+    rm key
+    set -x
+fi
