+++ fog05-plugin-fdu-kvm/generate_deb.sh
@@ -2,6 +2,13 @@
 
 set -e
 
+
+if [[ -z "${DEPLOY}" ]]; then
+  UPLOAD=false
+else
+  UPLOAD=true
+fi
+
 UBUNTU="ubuntu:bionic"
 DEBIAN="debian:10-slim"
 
@@ -38,10 +45,11 @@ docker cp build-kvm:/root/build/fog05-plugin-fdu-kvm_${VERSION}-1_amd64.deb ../f
 
 docker container rm --force build-kvm
 
-
-set +x
-echo $KEY  | base64 --decode > key
-chmod 0600 key
-scp -o StrictHostKeyChecking=no -i ./key ../fog05-plugin-fdu-kvm_${VERSION}-1_amd64.deb $USER@$SERVER:$DEPLOYDIR/fos/deb/bionic/amd64/fog05-plugin-fdu-kvm-1_amd64.deb
-rm key
-set -x
+if ["$UPLOAD" = true ]; then
+    set +x
+    echo $KEY  | base64 --decode > key
+    chmod 0600 key
+    scp -o StrictHostKeyChecking=no -i ./key ../fog05-plugin-fdu-kvm_${VERSION}-1_amd64.deb $USER@$SERVER:$DEPLOYDIR/fos/deb/bionic/amd64/fog05-plugin-fdu-kvm-1_amd64.deb
+    rm key
+    set -x
+fi
