+++ fog05-plugin-fdu-ros2/generate_deb.sh
@@ -2,6 +2,12 @@
 
 set -e
 
+if [[ -z "${DEPLOY}" ]]; then
+  UPLOAD=false
+else
+  UPLOAD=true
+fi
+
 UBUNTU="ubuntu:bionic"
 DEBIAN="debian:10-slim"
 
@@ -36,10 +42,11 @@ docker cp build-ros:/root/build/fog05-plugin-fdu-ros2_${VERSION}-1_amd64.deb ../
 docker container rm --force build-ros
 
 
-
-set +x
-echo $KEY  | base64 --decode > key
-chmod 0600 key
-scp -o StrictHostKeyChecking=no -i ./key ../fog05-plugin-fdu-ros2_${VERSION}-1_amd64.deb $USER@$SERVER:$DEPLOYDIR/fos/deb/bionic/amd64/fog05-plugin-fdu-ros2-1_amd64.deb
-rm key
-set -x
+if ["$UPLOAD" = true ]; then
+    set +x
+    echo $KEY  | base64 --decode > key
+    chmod 0600 key
+    scp -o StrictHostKeyChecking=no -i ./key ../fog05-plugin-fdu-ros2_${VERSION}-1_amd64.deb $USER@$SERVER:$DEPLOYDIR/fos/deb/bionic/amd64/fog05-plugin-fdu-ros2-1_amd64.deb
+    rm key
+    set -x
+fi
