+++ zenoh/generate_deb.sh
@@ -2,6 +2,12 @@
 
 set -e
 
+if [[ -z "${DEPLOY}" ]]; then
+  UPLOAD=false
+else
+  UPLOAD=true
+fi
+
 docker pull ${IMAGE}
 docker run -it -d --name build-z ${IMAGE} bash
 
@@ -55,10 +61,11 @@ docker cp build-z:/root/zenoh-${VERSION}.tar.gz ../zenoh-${VERSION}.tar.gz
 docker container rm --force build-z
 
 
-
-set +x
-echo $KEY  | base64 --decode > key
-chmod 0600 key
-scp -o StrictHostKeyChecking=no -i ./key ../zenoh_${VERSION}-1_amd64.deb $USER@$SERVER:$DEPLOYDIR/fos/deb/bionic/amd64/zenoh-1_amd64.deb
-rm key
-set -x
+if ["$UPLOAD" = true ]; then
+    set +x
+    echo $KEY  | base64 --decode > key
+    chmod 0600 key
+    scp -o StrictHostKeyChecking=no -i ./key ../zenoh_${VERSION}-1_amd64.deb $USER@$SERVER:$DEPLOYDIR/fos/deb/bionic/amd64/zenoh-1_amd64.deb
+    rm key
+    set -x
+fi
